Revert "Change HIDL fmq tests to use int32_t for payload"
Revert "Change payload type to int32_t for fmq device tests"

Revert submission 1431616-thirtytwo

Reason for revert: This topic broke the aosp_x86-userdebug target. Details at http://b/169077648

Reverted Changes:
If736389eb:Add template parameters to MQDescriptor
Ice9c28c3d:Change HIDL fmq tests to use int32_t for payload
Ia18d8459b:Add template arguments to MQDescriptor for AIDL
I95586cc83:Change payload type to int32_t for fmq device test...

+++ tests/msgq/1.0/default/TestMsgQ.h
@@ -49,13 +49,13 @@ using android::hardware::MQDescriptorUnsync;
 using android::hardware::MessageQueue;
 
 struct TestMsgQ : public ITestMsgQ {
-    typedef MessageQueue<int32_t, kSynchronizedReadWrite> MessageQueueSync;
-    typedef MessageQueue<int32_t, kUnsynchronizedWrite> MessageQueueUnsync;
+    typedef MessageQueue<uint16_t, kSynchronizedReadWrite> MessageQueueSync;
+    typedef MessageQueue<uint16_t, kUnsynchronizedWrite> MessageQueueUnsync;
 
     TestMsgQ() : mFmqSynchronized(nullptr), mFmqUnsynchronized(nullptr) {}
 
     // Methods from ::android::hardware::tests::msgq::V1_0::ITestMsgQ follow.
-    Return<bool> configureFmqSyncReadWrite(const MQDescriptorSync<int32_t>& mqDesc) override;
+    Return<bool> configureFmqSyncReadWrite(const MQDescriptorSync<uint16_t>& mqDesc) override;
     Return<void> getFmqUnsyncWrite(bool configureFmq, getFmqUnsyncWrite_cb _hidl_cb) override;
     Return<bool> requestWriteFmqSync(int32_t count) override;
     Return<bool> requestReadFmqSync(int32_t count) override;
@@ -73,7 +73,7 @@ private:
     /*
      * Utility function to verify data read from the fast message queue.
      */
-    bool verifyData(int32_t* data, int count) {
+    bool verifyData(uint16_t* data, int count) {
         for (int i = 0; i < count; i++) {
             if (data[i] != i) return false;
         }
